refactor: moved store hooks to folders
