+++ apps/davi/src/components/ActionsBuilder/SupportedActions/SetGuildConfig/validateSetGuildConfig.ts
@@ -2,7 +2,7 @@ import { TFunction } from 'react-i18next';
 import { SetGuildConfigFields } from './types';
 import { bn } from 'utils/safeBn';
 import { removeNullValues } from 'utils/object';
-import { GuildConfigProps } from 'stores/modules/common/fetchers/rpc/useGuildConfig';
+import { GuildConfigProps } from 'types/types.guilds.d';
 
 interface Context {
   t: TFunction;
