+++ apps/davi/src/stores/modules/common/fetchers/rpc/useDAOToken/index.ts
@@ -1,0 +1,1 @@
+export { useDAOToken } from './useDAOToken';
