+++ apps/davi/src/stores/modules/common/fetchers/rpc/useGetAllPermissions/index.ts
@@ -1,0 +1,1 @@
+export { useGetAllPermissions } from './useGetAllPermissions';
