+++ apps/davi/src/stores/modules/common/fetchers/rpc/useGetAllPermissions/useGetAllPermissions.tsx
