+++ apps/davi/src/stores/modules/common/fetchers/rpc/useGetMemberList/index.ts
@@ -1,0 +1,1 @@
+export { useGetMemberList } from './useGetMemberList';
