+++ apps/davi/src/stores/modules/common/fetchers/rpc/useGetNumberOfActiveProposals/index.ts
@@ -1,0 +1,1 @@
+export { useGetNumberOfActiveProposals } from './useGetNumberOfActiveProposals';
