+++ apps/davi/src/stores/modules/common/fetchers/rpc/useGetPermissions/index.ts
@@ -1,0 +1,1 @@
+export { useGetPermissions } from './useGetPermissions';
