+++ apps/davi/src/stores/modules/common/fetchers/rpc/useGetPermissions/useGetPermissions.tsx
@@ -2,9 +2,9 @@ import { useContractRead } from 'wagmi';
 import { Permission } from 'components/ActionsBuilder/types';
 import { PermissionRegistry } from 'contracts/ts-files/PermissionRegistry';
 import { FetcherHooksInterface } from 'stores/types';
-import { useListenToPermissionSet } from '../../events';
+import { useListenToPermissionSet } from '../../../events';
 import { ZERO_ADDRESS, ZERO_FUNC_SIGNATURE } from 'utils';
-import { useGuildConfig } from '.';
+import { useGuildConfig } from '..';
 
 type IUseGetPermissions = FetcherHooksInterface['useGetPermissions'];
