+++ apps/davi/src/stores/modules/common/fetchers/rpc/useGetVotes/index.ts
@@ -1,0 +1,1 @@
+export { useGetVotes } from './useGetVotes';
