+++ apps/davi/src/stores/modules/common/fetchers/rpc/useGuildConfig/index.ts
@@ -1,0 +1,1 @@
+export { useGuildConfig } from './useGuildConfig';
