+++ apps/davi/src/stores/modules/common/fetchers/rpc/useGuildProposalIds/index.ts
@@ -1,0 +1,1 @@
+export { useGuildProposalIds } from './useGuildProposalIds';
