+++ apps/davi/src/stores/modules/common/fetchers/rpc/useIsProposalCreationAllowed/index.ts
@@ -1,0 +1,1 @@
+export { useIsProposalCreationAllowed } from './useIsProposalCreationAllowed';
