+++ apps/davi/src/stores/modules/common/fetchers/rpc/useIsProposalCreationAllowed/useIsProposalCreationAllowed.tsx
