+++ apps/davi/src/stores/modules/common/fetchers/rpc/useProposal/index.ts
@@ -1,0 +1,1 @@
+export { useProposal } from './useProposal';
