+++ apps/davi/src/stores/modules/common/fetchers/rpc/useProposalCalls/index.ts
@@ -1,0 +1,1 @@
+export { useProposalCalls } from './useProposalCalls';
