+++ apps/davi/src/stores/modules/common/fetchers/rpc/useProposalVotesOfVoter/index.ts
@@ -1,0 +1,1 @@
+export { useProposalVotesOfVoter } from './useProposalVotesOfVoter';
