+++ apps/davi/src/stores/modules/common/fetchers/rpc/useSnapshotId/index.ts
@@ -1,0 +1,1 @@
+export { useSnapshotId } from './useSnapshotId';
