+++ apps/davi/src/stores/modules/common/fetchers/rpc/useVotingResults/index.ts
@@ -1,0 +1,1 @@
+export { useVotingResults } from './useVotingResults';
