+++ apps/davi/src/stores/modules/common/fetchers/rpc/useVotingResults/useVotingResults.ts
